Apps/requirements.txt
Django==2.2.2
django-phone-field==1.7.1
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
